31.05 restore point

+++ src/main/java/lemon/hospitaltable/table/objects/Patient.java
@@ -5,6 +5,6 @@ import org.springframework.data.relational.core.mapping.Table;
 import java.sql.Date;
 
 @Table("patients")
-public record Patient(@Id Integer id, String name, Date birth, String address, String status, String notation) {
+public record Patient(@Id Long id, String name, Date birth, String address, String status, String notation) {
 
 }
