+++ src/main/java/lemon/hospitaltable/table/objects/Treatment.java
@@ -6,7 +6,7 @@ import org.springframework.data.relational.core.mapping.Table;
 import java.sql.Date;
 
 @Table("treatments")
-public record Treatment(@Id Integer id, @Column("patient_id") Integer patientId, @Column("doctor_id") Integer doctorId,
+public record Treatment(@Id Long id, @Column("patient_id") Long patientId, @Column("doctor_id") Integer doctorId,
                         @Column("ward_id") Integer wardId, Date dateIn, Date dateOut, String notation) {
 
 }
